TP2/requirements.txt
aiohttp
beautifulsoup4
lxml
Pillow
playwright